Implented ActionPacketQueue / UI Fixes
Drag drop actions and some  use object actions will be enqueued on a seperate thread that will pause the remaining ActionDelayMS between actions, this should reduce the likelyhood of 'You must wait to perform...' when using Dress/Loot/Scavenger agents.
Fixed UI issues and increased the size of the ClilocFilterConfigureWindow.