Asynchronous PlaySound option added.
Added an optional second argument to PlaySound() that pauses the script for a designated period of time rather than until the sound finishes playing.  This is useful both to allow a script to continue executing while a sound is played and to easily add a longer delay between looped sounds.