Autoloot/ECV Filter 'ID (Multiple)' constraint (#521)
* Refactors Autoloot property loading.

Moves Autoloot property loading and assembly loading logic to AutolootManager.
This change centralizes the property management, and removes redundant loading logic from the ECV and Autoloot View Models.

* Null check Assemblies

* Autoloot/ECV Filter 'ID (Multiple)' constraint

* Nitpicks

* Simplify tests to Match

* Copilot Fixes

* Filters autoloot operators based on property

Adds functionality to filter the available autoloot operators
in the UI based on the selected property, improving usability
and preventing invalid configurations.

Introduces `AutolootAllowedOperators` enum to define allowed operators
for each property and a value converter to filter the operator list
in the UI.

* Fix

* Fix tests